requirements.txt
# Core Dependencies
pandas>=2.0.0
openpyxl>=3.1.0
pillow>=10.0.0
requests>=2.31.0
pyyaml>=6.0
openai>=1.0.0

# Progress & Logging
tqdm>=4.66.0
colorama>=0.4.6

# Optional
python-dotenv>=1.0.0